# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  choices.AccType:
    enum:
    - READER
    - AUTHOR
    type: string
    x-enum-varnames:
    - ACCTYPE_READER
    - ACCTYPE_AUTHOR
...
  choices.NotificationTypeChoice:
    enum:
    - LIKE
    - REPLY
    - FOLLOWING
    - BOOK_PURCHASE
    - GIFT
    - REVIEW
    - VOTE
    type: string
    x-enum-varnames:
    - NT_LIKE
    - NT_REPLY
    - NT_FOLLOWING
    - NT_BOOK_PURCHASE
    - NT_GIFT
    - NT_REVIEW
    - NT_VOTE
...
  choices.SubscriptionTypeChoice:
    enum:
    - MONTHLY
    - ANNUAL
    type: string
    x-enum-varnames:
    - ST_MONTHLY
    - ST_ANNUAL
...
  schemas.FollowerData:
    properties:
      account_type:
        $ref: '#/definitions/choices.AccType'
      avatar:
        type: string
      followers_count:
        type: integer
      name:
        type: string
      stories_count:
        type: integer
      username:
        type: string
    type: object
...
  schemas.NotificationBookSchema:
    properties:
      cover_image:
        type: string
      slug:
        type: string
      title:
        type: string
    type: object
  schemas.NotificationSchema:
    properties:
      book:
        allOf:
        - $ref: '#/definitions/schemas.NotificationBookSchema'
        description: Bought book, vote, comment and reply
      created_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      id:
        example: 2b3bd817-135e-41bd-9781-33807c92ff40
        type: string
      is_read:
        type: boolean
      ntype:
        $ref: '#/definitions/choices.NotificationTypeChoice'
      receiver_id:
        type: string
      reply_id:
        description: If someone liked your reply
        example: 2b3bd817-135e-41bd-9781-33807c92ff40
        type: string
      review_id:
        description: reviewed, reply, like
        example: 2b3bd817-135e-41bd-9781-33807c92ff40
        type: string
      sender:
        $ref: '#/definitions/schemas.UserDataSchema'
      sent_gift_id:
        description: If someone sent you a gift
        example: 2b3bd817-135e-41bd-9781-33807c92ff40
        type: string
      text:
        type: string
    type: object
  schemas.NotificationsResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      notifications:
        items:
          $ref: '#/definitions/schemas.NotificationSchema'
        type: array
      per_page:
        example: 100
        type: integer
    type: object
  schemas.NotificationsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.NotificationsResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.ReadNotificationSchema:
    properties:
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      mark_all_as_read:
        example: false
        type: boolean
    type: object
...
  schemas.ResponseSchema:
    properties:
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.UpdatePasswordSchema:
    properties:
      new_password:
        example: oldpassword
        maxLength: 50
        minLength: 8
        type: string
      old_password:
        example: newstrongpassword
        maxLength: 50
        minLength: 8
        type: string
    required:
    - new_password
    - old_password
    type: object
  schemas.UpdateUserRoleSchema:
    properties:
      account_type:
        allOf:
        - $ref: '#/definitions/choices.AccType'
        example: WRITER
    type: object
  schemas.UserDataSchema:
    properties:
      avatar:
        type: string
      name:
        description: For short user data
        type: string
      username:
        type: string
    type: object
...
  schemas.UserProfile:
    properties:
      account_type:
        $ref: '#/definitions/choices.AccType'
      avatar:
        type: string
      bio:
        type: string
      created_at:
        example: "2024-06-05T02:32:34.462196+01:00"
        type: string
      current_plan:
        $ref: '#/definitions/choices.SubscriptionTypeChoice'
      email:
        type: string
      followers:
        items:
          $ref: '#/definitions/schemas.FollowerData'
        type: array
      followings:
        items:
          $ref: '#/definitions/schemas.FollowerData'
        type: array
      name:
        type: string
      stories_count:
        type: integer
      username:
        type: string
    type: object
  schemas.UserProfileResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.UserProfile'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.UserProfilesResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
      users:
        items:
          $ref: '#/definitions/schemas.UserProfile'
        type: array
    type: object
  schemas.UserProfilesResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.UserProfilesResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      code:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: |-
    ## LitPAD API built with Fiber and GORM

    ### WEBSOCKETS:

    #### Notifications

    - URL: `wss://{host}/api/v1/ws/notifications`

    - Requires authorization, so pass in the Bearer Authorization header.

    - You can only read and not send notification messages into this socket.
  title: LITPAD API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retrieves a list of user profiles with support for pagination and
        optional filtering based on user account type.
      parameters:
      - description: Type of user to filter by
        enum:
        - READER
        - WRITER
        - ADMIN
        in: query
        name: account_type
        type: string
      - default: 1
        description: Current page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of user profiles
          schema:
            $ref: '#/definitions/schemas.UserProfilesResponseSchema'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Users with Pagination
      tags:
      - Admin | Users
  /admin/users/{username}:
    put:
      consumes:
      - application/json
      description: Updates the account type of a specified user.
      parameters:
      - default: username
        description: Username
        in: path
        name: username
        required: true
        type: string
      - description: User role update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateUserRoleSchema'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user details
          schema:
            $ref: '#/definitions/schemas.UserProfileResponseSchema'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User Role
      tags:
      - Admin | Users
  /admin/users/{username}/toggle-activation:
    get:
      consumes:
      - application/json
      description: Allows the admin to deactivate/reactivate a user.
      parameters:
      - default: username
        description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reactivate/Deactivate User
      tags:
      - Admin | Users
...
  /profiles/notifications:
    get:
      description: This endpoint allows a user to view his/her notificatios
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.NotificationsResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: View Notifications
      tags:
      - Profiles
  /profiles/notifications/read:
    post:
      description: This endpoint allows a user to read his/her notification.
      parameters:
      - description: Notification Read object
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/schemas.ReadNotificationSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Read Notification
      tags:
      - Profiles
  /profiles/profile/{username}:
    get:
      description: This endpoint views a user profile
      parameters:
      - description: Username of user
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserProfileResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: View User Profile
      tags:
      - Profiles
  /profiles/profile/{username}/follow:
    get:
      description: '`This endpoint allows a user to follow or unfollow a writer`.'
      parameters:
      - description: Username of the user to follow or unfollow
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Returns an error for invalid request parameters.
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Returns an error when trying to follow a user that isn't a
            writer
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Returns an error when either the user to follow or the follower
            user does not exist.
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Returns an error when there is an internal server error or
            a transaction fails.
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle Follow Status
      tags:
      - Profiles
  /profiles/update:
    patch:
      description: This endpoint updates a user's profile
      parameters:
      - example: I'm here to read good books
        in: formData
        maxLength: 1000
        minLength: 3
        name: bio
        type: string
      - example: John Doe
        in: formData
        maxLength: 1000
        minLength: 3
        name: name
        type: string
      - example: johndoe
        in: formData
        maxLength: 1000
        minLength: 3
        name: username
        type: string
      - description: Avatar Image to upload
        in: formData
        name: avatar
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserProfileResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - Profiles
  /profiles/update-password:
    put:
      description: This endpoint updates a user's password
      parameters:
      - description: Password object
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdatePasswordSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User Password
      tags:
      - Profiles
...
produces:
- application/json
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer jwt_string' to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"